\name{evdbayes-internal}

\alias{dprior.quant}
\alias{dprior.prob}
\alias{dprior.norm}
\alias{dprior.loglognorm}
\alias{gevlik2}
\alias{gpdlik2}
\alias{pplik2}
\alias{oslik2}
\alias{dpost}
\alias{gibbs}

\title{Internal Functions}

\description{
  Not to be called by the user.
}

\value{
  Internal functions, returning numeric vectors. The only exception is the internal function
  \code{gibbs} which is used to produce the output of \code{\link{posterior}}.
}

\keyword{internal}
