\name{rainfall}
\alias{rainfall}
\title{Daily Aggregate Rainfall}
\usage{data(rainfall)}
\description{
  A numeric vector of length 20820 containing daily aggregate
  rainfall observations, in millimetres, recorded at a rain gauge
  in England over a 57 year period, beginning on a leap year.
  Three of these years contain only missing values.
}
\format{A vector containing 20820 observations.}
\source{
  Unknown.
}  
\keyword{datasets}
