\name{evdbayes-internal}

\alias{dprior.quant}
\alias{dprior.prob}
\alias{dprior.norm}
\alias{dprior.loglognorm}
\alias{gevlik2}
\alias{gpdlik2}
\alias{pplik2}
\alias{oslik2}
\alias{dpost}
\alias{gibbs}
\alias{gibbs.mix}

\title{Internal Functions}

\description{
  Not to be called by the user.
}

\keyword{internal}
